// src/lib/CIFTools.d.ts
    const enum ValuePresence {
        Present = 0,
        NotSpecified = 1,
        Unknown = 2,
    }
// ...
    interface FieldDesc<Data> {
        name: string;
        string?: (data: Data, i: number) => string | null;
        number?: (data: Data, i: number) => number;
        typedArray?: any;
        encoder?: Binary.Encoder;
        presence?: (data: Data, i: number) => ValuePresence;
    }
    interface CategoryDesc<Data> {
        name: string;
        fields: FieldDesc<Data>[];
    }
    type CategoryInstance<Data> = {
        data: any;
        count: number;
        desc: CategoryDesc<Data>;
    };
    type CategoryProvider = (ctx: any) => CategoryInstance<any> | undefined;
    type OutputStream = {
        writeString: (data: string) => boolean;
        writeBinary: (data: Uint8Array) => boolean;
    };
    interface Writer<Context> {
        startDataBlock(header: string): void;
        writeCategory(category: CategoryProvider, contexts?: Context[]): void;
        encode(): void;
        flush(stream: OutputStream): void;
    }